refactor: refactor workflow-approval plugin
