+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/config/LauncherActionConfig.tsx
@@ -1,6 +1,6 @@
 import { SchemaInitializer } from '@tachybase/client';
 
-import { APPROVAL_ACTION_STATUS } from '../../constants';
+import { APPROVAL_INITIATION_STATUS } from '../../../../common/constants/approval-initiation-status';
 import { NAMESPACE } from '../../locale';
 import { LauncherActionConfigComponent } from './LauncherActionConfigComponent';
 
@@ -14,7 +14,7 @@ export const LauncherActionConfigInitializer = new SchemaInitializer({
       type: 'item',
       title: '{{t("Submit")}}',
       Component: LauncherActionConfigComponent,
-      action: APPROVAL_ACTION_STATUS.SUBMITTED,
+      action: APPROVAL_INITIATION_STATUS.SUBMITTED,
       actionProps: { type: 'primary' },
       disabled: true,
     },
@@ -23,7 +23,7 @@ export const LauncherActionConfigInitializer = new SchemaInitializer({
       type: 'item',
       title: `{{t("Save draft", { ns: "${NAMESPACE}" })}}`,
       Component: LauncherActionConfigComponent,
-      action: APPROVAL_ACTION_STATUS.DRAFT,
+      action: APPROVAL_INITIATION_STATUS.DRAFT,
     },
   ],
 });
