+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/provider/ActionBar.tsx
@@ -1,6 +1,5 @@
 import { useCurrentUserContext } from '@tachybase/client';
 
-import { APPROVAL_ACTION_STATUS } from '../../constants';
 import { useContextApprovalExecution } from '../../context/ApprovalExecution';
 
 export function ActionBarProvider(props) {
