+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/provider/ActionReminder.provider.tsx
@@ -1,6 +1,6 @@
 import { useCurrentUserContext } from '@tachybase/client';
 
-import { APPROVAL_ACTION_STATUS } from '../../constants';
+import { APPROVAL_INITIATION_STATUS } from '../../../../common/constants/approval-initiation-status';
 import { useContextApprovalExecution } from '../../context/ApprovalExecution';
 
 export const ProviderActionReminder = (props) => {
@@ -11,8 +11,8 @@ export const ProviderActionReminder = (props) => {
 
   const isSameId = data.data.id === createdById;
   const isSameExecutionId = latestExecutionId === id;
-  const isDraft = status === APPROVAL_ACTION_STATUS.DRAFT;
-  const isReturned = status === APPROVAL_ACTION_STATUS.RETURNED;
+  const isDraft = status === APPROVAL_INITIATION_STATUS.DRAFT;
+  const isReturned = status === APPROVAL_INITIATION_STATUS.RETURNED;
 
   if (isSameId && !isDraft && !isReturned && isSameExecutionId) {
     return props.children;
