+++ packages/plugin-workflow-approval/src/client/usage/h5/initiations/provider/ActionResubmit.provider.tsx
@@ -1,6 +1,6 @@
 import { useCurrentUserContext } from '@tachybase/client';
 
-import { APPROVAL_ACTION_STATUS } from '../../constants';
+import { APPROVAL_INITIATION_STATUS } from '../../../../common/constants/approval-initiation-status';
 import { useContextApprovalExecution } from '../../context/ApprovalExecution';
 import { useResubmit } from './Resubmit.provider';
 
@@ -11,8 +11,8 @@ export function ProviderActionResubmit(props) {
   const { status, createdById } = approval;
 
   const isSameId = data.data.id === createdById;
-  const isDraft = status === APPROVAL_ACTION_STATUS.DRAFT;
-  const isReturned = status === APPROVAL_ACTION_STATUS.RETURNED;
+  const isDraft = status === APPROVAL_INITIATION_STATUS.DRAFT;
+  const isReturned = status === APPROVAL_INITIATION_STATUS.RETURNED;
   if (isSameId && !isResubmit && !isDraft && !isReturned) {
     return props.children;
   }
